fix(models): Change ActionLink.id from i32 to i64
The DigitalOcean API returns action IDs that exceed i32::MAX (e.g., 2931943763).
This fix changes the id field from Option<i32> to Option<i64> to properly
deserialize these values.

Also adds a patch file to preserve this fix across code regeneration cycles.

+++ src/models/action_link.rs
@@ -16,7 +16,7 @@ use serde::{Deserialize, Serialize};
 pub struct ActionLink {
     /// A unique numeric ID that can be used to identify and reference an action.
     #[serde(rename = "id", skip_serializing_if = "Option::is_none")]
-    pub id: Option<i32>,
+    pub id: Option<i64>,
     /// A string specifying the type of the related action.
     #[serde(rename = "rel", skip_serializing_if = "Option::is_none")]
     pub rel: Option<String>,
